// src/app/pages/contract/enums/contract-data.enum.ts
export enum ContractDataEnum {
    VND = 'VND',
    USD = 'USD',
    Day = 'day'
  }
  export enum ContractDepositTypeEnum {
    BankGuarantee = 'BankGuarantee',
    Deposit = 'Deposit',
    Prepaid = 'Prepaid'
  }
  export enum ContractCurrencyExchangeEnum {
    Fix = 'Fix',
    Quotation = 'Quotation',
    BillingDate = 'BillingDate'
  }
  export enum ContractAnnexType {
    ScopeOfWork = 'ScopeOfWork',
    Quotation = 'Quotation',
    Billing = 'Billing',
    OnBehalf = 'OnBehalf',
    Deposit = 'Deposit',
    ServiceCharges = 'ServiceCharges',
    ExtendContractDuration = 'ExtendContractDuration'
  }
  